repl: enable experimental repl
