+++ doc/node.1
@@ -182,6 +182,9 @@ Enable the experimental permission model.
 .It Fl -experimental-shadow-realm
 Use this flag to enable ShadowRealm support.
 .
+.It Fl -experimental-repl
+Enable the Experimental REPL.
+.
 .It Fl -experimental-test-coverage
 Enable code coverage in the test runner.
 .
