+++ lib/internal/main/repl.js
@@ -36,9 +36,6 @@ if (process.env.NODE_REPL_EXTERNAL_MODULE) {
   }
 
   require('internal/modules/run_main').runEntryPointWithESMLoader(() => {
-    console.log(`Welcome to Node.js ${process.version}.\n` +
-      'Type ".help" for more information.');
-
     const cliRepl = require('internal/repl');
     cliRepl.createInternalRepl(process.env, (err, repl) => {
       if (err) {
