+++ lib/internal/repl.js
@@ -6,13 +6,22 @@ const {
   NumberParseInt,
 } = primordials;
 
-const REPL = require('repl');
-const { kStandaloneREPL } = require('internal/repl/utils');
+const StableREPL = require('internal/repl/stable/index');
+const ExperimentalREPL = require('internal/repl/experimental/index');
+const { kStandaloneREPL } = require('internal/repl/stable/utils');
+const {
+  getOptionValue,
+} = require('internal/options');
 
-module.exports = { __proto__: REPL };
-module.exports.createInternalRepl = createRepl;
+const useExperimentalRepl = getOptionValue('--experimental-repl') || false;
 
-function createRepl(env, opts, cb) {
+module.exports = { __proto__: useExperimentalRepl ? ExperimentalREPL : StableREPL };
+module.exports.createInternalRepl = useExperimentalRepl ? createExperimentalRepl : createStableRepl;
+
+function createStableRepl(env, opts, cb) {
+  process.stdout.write(`Welcome to Node.js ${process.version}.\n` +
+      'Type ".help" for more information.\n');
+
   if (typeof opts === 'function') {
     cb = opts;
     opts = null;
@@ -31,13 +40,13 @@ function createRepl(env, opts, cb) {
 
   if (env.NODE_REPL_MODE) {
     opts.replMode = {
-      'strict': REPL.REPL_MODE_STRICT,
-      'sloppy': REPL.REPL_MODE_SLOPPY,
+      'strict': StableREPL.REPL_MODE_STRICT,
+      'sloppy': StableREPL.REPL_MODE_SLOPPY,
     }[env.NODE_REPL_MODE.toLowerCase().trim()];
   }
 
   if (opts.replMode === undefined) {
-    opts.replMode = REPL.REPL_MODE_SLOPPY;
+    opts.replMode = StableREPL.REPL_MODE_SLOPPY;
   }
 
   const historySize = Number(env.NODE_REPL_HISTORY_SIZE);
@@ -47,7 +56,13 @@ function createRepl(env, opts, cb) {
     opts.historySize = 1000;
   }
 
-  const repl = REPL.start(opts);
+  const repl = StableREPL.start(opts);
   const term = 'terminal' in opts ? opts.terminal : process.stdout.isTTY;
   repl.setupHistory(term ? env.NODE_REPL_HISTORY : '', cb);
+}
+
+async function createExperimentalRepl() {
+  const repl = new ExperimentalREPL();
+  await repl.start();
+  ExperimentalREPL.repl = repl; // For introspection
 }
