+++ lib/internal/repl/stable/await.js
@@ -20,7 +20,7 @@ const {
 
 const parser = require('internal/deps/acorn/acorn/dist/acorn').Parser;
 const walk = require('internal/deps/acorn/acorn-walk/dist/walk');
-const { Recoverable } = require('internal/repl');
+const { Recoverable } = require('internal/repl/stable/index');
 
 function isTopLevelDeclaration(state) {
   return state.ancestors[state.ancestors.length - 2] === state.body;
