+++ lib/internal/repl/stable/utils.js
